chat atomic desgin
