+++ src/components/Molecules/Chats/JobsLabel.vue
@@ -1,0 +1,16 @@
+<script>
+export default {
+
+}
+</script>
+
+<template>
+  <section>
+    aaaa
+  </section>
+</template>
+
+
+<style>
+
+</style>
